pyproject.toml
[project]
name = "persona_steering"
version = "0.1.0"
description = "Layer-wise persona vector steering experiments"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
  # --- Core ---
  "numpy>=1.24",
  "scipy>=1.10",
  "tqdm>=4.66",
  "pyyaml>=6.0",
  "regex>=2023.12.25",
  # --- Hugging Face / LLM ---
  "transformers>=4.41",
  "accelerate>=0.31",
  "tokenizers>=0.19",
  "safetensors>=0.4",
  "datasets>=2.19",
  "huggingface-hub>=0.23",
  # --- Metrics / evaluation ---
  "scikit-learn>=1.3",
  "sentence-transformers>=3.0",
  "torch>=2.9.1",
  "pyarrow>=22.0.0",
  "fsspec>=2025.10.0",
  "sentencepiece>=0.2.1",
  "protobuf>=6.33.2",
  "pandas>=2.3.3",
  "matplotlib>=3.10.8",
  "bitsandbytes>=0.49.0",
  "einops>=0.8.1",
]

[tool.uv]
# uvはlockを自動生成してくれます（uv.lock）
